requirements.txt
crewai
crewai-tools
pydantic
tavily-python
scrapegraph-py
python-dotenv
